2.0.0
The interace for retreiving and decoding attributes has changed. This
change was made in response both to my own usability frustrations and
feedback from consumers of the library.

+ Rather than being a Dict, attributes are structured as a raw Json
value.
+ The "attributes" function now takes a decoder as an argument. The
decoder should parse the Json Value of the attributes.
+ I have also added an "attribute" function for those interested in
retreiving a single value. This function takes a string (the attribute
name) in addition to a Json decoder.